# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabeleleila Leila",
    "description": "Sistema de agendamento para o salão Cabeleleila Leila",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/appointments": {
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Create Appointment",
        "description": "Solicita um novo agendamento.",
        "operationId": "create_appointment_api_appointments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "List Appointments",
        "description": "Lista agendamentos (admin: todos, cliente: seus).",
        "operationId": "list_appointments_api_appointments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Data início (YYYY-MM-DD)",
              "title": "Date From"
            },
            "description": "Data início (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Data fim (YYYY-MM-DD)",
              "title": "Date To"
            },
            "description": "Data fim (YYYY-MM-DD)"
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "ID do cliente (admin apenas)",
              "title": "Client Id"
            },
            "description": "ID do cliente (admin apenas)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/check-week": {
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Check Week Appointments",
        "description": "Verifica se o cliente já possui agendamentos na mesma semana da data informada.",
        "operationId": "check_week_appointments_api_appointments_check_week_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data no formato YYYY-MM-DD",
              "title": "Date"
            },
            "description": "Data no formato YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AppointmentResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Check Week Appointments Api Appointments Check Week Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/available-slots": {
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Get Available Slots",
        "description": "Retorna horários disponíveis para uma data.",
        "operationId": "get_available_slots_api_appointments_available_slots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data no formato YYYY-MM-DD",
              "title": "Date"
            },
            "description": "Data no formato YYYY-MM-DD"
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Duração total dos serviços em minutos",
              "default": 15,
              "title": "Duration"
            },
            "description": "Duração total dos serviços em minutos"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableSlotResponse"
                  },
                  "title": "Response Get Available Slots Api Appointments Available Slots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments/revenue-chart": {
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Get Revenue Chart",
        "description": "Retorna dados de receita para o gráfico (admin apenas).",
        "operationId": "get_revenue_chart_api_appointments_revenue_chart_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "monthly ou daily",
              "default": "monthly",
              "title": "Period"
            },
            "description": "monthly ou daily"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Data início (YYYY-MM-DD)",
              "title": "Date From"
            },
            "description": "Data início (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Data fim (YYYY-MM-DD)",
              "title": "Date To"
            },
            "description": "Data fim (YYYY-MM-DD)"
          },
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "ID do serviço para filtro",
              "title": "Service Id"
            },
            "description": "ID do serviço para filtro"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Status do agendamento para filtro",
              "title": "Status"
            },
            "description": "Status do agendamento para filtro"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevenueChartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{appointment_id}": {
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Get Appointment",
        "description": "Busca um agendamento por ID.",
        "operationId": "get_appointment_api_appointments__appointment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Update Appointment",
        "description": "Edita um agendamento (respeita regra dos 2 dias para clientes).",
        "operationId": "update_appointment_api_appointments__appointment_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{appointment_id}/approve": {
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Approve Appointment",
        "description": "Aprova um agendamento pendente (admin apenas).",
        "operationId": "approve_appointment_api_appointments__appointment_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{appointment_id}/reject": {
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Reject Appointment",
        "description": "Rejeita um agendamento pendente (admin apenas).",
        "operationId": "reject_appointment_api_appointments__appointment_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{appointment_id}/cancel": {
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Cancel Appointment",
        "description": "Cancela um agendamento.",
        "operationId": "cancel_appointment_api_appointments__appointment_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{appointment_id}/complete": {
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Complete Appointment",
        "description": "Marca agendamento como concluído (admin apenas).",
        "operationId": "complete_appointment_api_appointments__appointment_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppointmentCreate": {
        "properties": {
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "service_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Service Ids"
          },
          "add_to_google_calendar": {
            "type": "boolean",
            "title": "Add To Google Calendar",
            "default": false
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          }
        },
        "type": "object",
        "required": [
          "data_hora",
          "service_ids"
        ],
        "title": "AppointmentCreate",
        "description": "Schema para criação de agendamento (solicitação)."
      },
      "AppointmentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppointmentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "AppointmentListResponse",
        "description": "Schema de resposta para lista paginada de agendamentos."
      },
      "AppointmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "criado_por": {
            "type": "integer",
            "title": "Criado Por"
          },
          "aprovado_por": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aprovado Por"
          },
          "data_aprovacao": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Aprovacao"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/AppointmentServiceResponse"
            },
            "type": "array",
            "title": "Services",
            "default": []
          },
          "client_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Nome"
          },
          "approver_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Nome"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "data_hora",
          "status",
          "criado_por",
          "criado_em",
          "atualizado_em"
        ],
        "title": "AppointmentResponse",
        "description": "Schema de resposta de agendamento."
      },
      "AppointmentServiceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_id"
        ],
        "title": "AppointmentServiceResponse",
        "description": "Schema de resposta de serviço vinculado ao agendamento."
      },
      "AppointmentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELLED",
          "COMPLETED"
        ],
        "title": "AppointmentStatus",
        "description": "Status possíveis de um agendamento."
      },
      "AppointmentUpdate": {
        "properties": {
          "data_hora": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Hora"
          },
          "service_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Ids"
          }
        },
        "type": "object",
        "title": "AppointmentUpdate",
        "description": "Schema para atualização de agendamento."
      },
      "AvailableSlotResponse": {
        "properties": {
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "disponivel": {
            "type": "boolean",
            "title": "Disponivel",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "data_hora"
        ],
        "title": "AvailableSlotResponse",
        "description": "Schema de resposta de horários disponíveis."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevenueChartItem": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "estimated_revenue": {
            "type": "number",
            "title": "Estimated Revenue"
          },
          "paid_revenue": {
            "type": "number",
            "title": "Paid Revenue"
          }
        },
        "type": "object",
        "required": [
          "label",
          "estimated_revenue",
          "paid_revenue"
        ],
        "title": "RevenueChartItem",
        "description": "Item de dado para o gráfico de receita."
      },
      "RevenueChartResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RevenueChartItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_estimated": {
            "type": "number",
            "title": "Total Estimated"
          },
          "total_paid": {
            "type": "number",
            "title": "Total Paid"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_estimated",
          "total_paid"
        ],
        "title": "RevenueChartResponse",
        "description": "Resposta com dados para o gráfico e sumários."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}